+++ Mariana/layers.py
@@ -128,6 +128,20 @@ class Layer_ABC(object) :
 		for init in self.initializations :
 			init.apply(self)
 	
+	def initParameter(self, parameter, value) :
+		"""Initialize a parameter, raise value error if already initialized"""
+		if not hasattr(self, parameter) or getattr(self, parameter) is None :
+			setattr(self, parameter, value)
+		else :
+			raise ValueError("Parameter '%s' or layer '%s' has already been initialized" % (parameter, self.name) )
+
+	def updateParameter(self, parameter, value) :
+		"""Update the value of an already initialized parameter. Raise value error if the parameter has not been initialized"""
+		if k not in self.getParameterDict().keys() :
+			raise ValueError("Parameter '%s' has not been initialized as parameter of layer '%s'" % (parameter, self.name) )
+		else :
+			setattr(self, parameter, value)
+
 	#theano hates abstract methods defined with a decorator
 	def _setOutputs(self) :
 		"""Defines the outputs and testOutputs of the layer before the application of the activation function. This function is called by _init() ans should be written in child."""